Docker based statically linked build (#1)
## What did I do?
I modified the Makefile to statically link against the cuda libraries. This made it necessary to explicitly link against standard/gcc libraries as well.
I also modified the Dockerfile with an export-stage to output the produced binary and serve as a reproducible build environment. I also allowed for external configuration of the targeted cuda and ubuntu versions.


## Why did I do this?
This allows for easy and reproducible builds and is prerequisite of including gpu-burn in the new stress test. Statically linking makes it nicely portable across servers without the need to install or bundle a runtime library.
This is a prerequisite for including gpu-burn in the stress test.


## Testing
I have tested the binary against the following combinations:
 - Driver Version: 460.91.03    CUDA Version: 11.2
 - Driver Version: 410.78       CUDA Version: 10.0
 - Driver Version: 465.19.01    CUDA Version: 11.3


## Anything else?
The build is run like this (resulting binary and compare.ptx are stored in the folder `./build`)
`DOCKER_BUILDKIT=1 docker build --target export-stage --output type=local,dest=./build .`