Docker based statically linked build for new HW
I modified the Makefile to statically link against the cuda libraries.
This made it necessary to explicitly link against standard/gcc libraries as well.

I also modified the Dockerfile with an exporter stage
to output the produced binary and serve as a reproducible build environment.

This allows for easy and reproducible builds and is prerequisite of including gpu-burn in the new stress test.
Statically linking makes it nicely portable across servers without the need to install or bundle a runtime library.
This is a prerequisite for including gpu-burn in the stress test.

 I have tested the binary against the following combinations:
  - Thinkmate Mega M3D 4.0.2 with Driver Version: 460.73.01    CUDA Version: 11.2
  - Dell Rack M3D 4.1.0 with Driver Version: 535.183.01        CUDA Version: 12.2

The build is run like this (resulting binary and compare.ptx are stored in the folder `./build`)
`DOCKER_BUILDKIT=1 docker build --target exporter --output type=local,dest=./build .`